# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnalystGPT Enterprise",
    "description": "Enterprise-grade analytics platform providing data upload, cleaning, quality assessment, analytics, reporting, AI-powered business insights, Power BI integration, REST APIs, and Streamlit dashboard visualisation.",
    "contact": {
      "name": "Amal Jose"
    },
    "license": {
      "name": "MIT"
    },
    "version": "12.0.0"
  },
  "paths": {
...
    "/api/ai/jobs/{job_id}": {
      "get": {
        "tags": [
          "AI Insights",
          "AI Insights"
        ],
        "summary": "Get AI Job Status & Report",
        "description": "Retrieve the lifecycle status, attempt count, and generated AI insight report for a given AI job ID.",
        "operationId": "get_ai_job_api_ai_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-User-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          },
          {
            "name": "X-User-Email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/jobs/latest/status": {
      "get": {
        "tags": [
          "AI Insights",
          "AI Insights"
        ],
        "summary": "Get Latest AI Job",
        "description": "Retrieve the latest AI generation job for the authenticated user.",
        "operationId": "get_latest_ai_job_api_ai_jobs_latest_status_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-User-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          },
          {
            "name": "X-User-Email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "AI Insights",
          "AI Insights"
        ],
        "summary": "Retry Failed AI Job",
        "description": "Manually trigger a retry for an AI job in FAILED state.",
        "operationId": "retry_ai_job_api_ai_jobs__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-User-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Name"
            }
          },
          {
            "name": "X-User-Email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIJobRetryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Unique identifier of the AI job."
          },
          "pipeline_run_id": {
            "type": "integer",
            "title": "Pipeline Run Id",
            "description": "Associated pipeline execution run ID."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current lifecycle status: PENDING, GENERATING, READY, FAILED."
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "AI LLM provider name (e.g. ollama)."
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Configured model name (e.g. gemma3:4b)."
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count",
            "description": "Number of execution attempts made so far.",
            "default": 0
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts",
            "description": "Maximum retry attempts allowed.",
            "default": 3
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Sanitized error description if failed."
          },
          "failure_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Category",
            "description": "Failure classification category if failed."
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Timestamp when the job was created."
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "Timestamp when the job execution started."
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "Timestamp when the job completed (READY or FAILED)."
          },
          "ai_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIReportResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Generated AI insight report when status is READY."
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "pipeline_run_id",
          "status",
          "provider",
          "model"
        ],
        "title": "AIJobResponse",
        "description": "Standard response model for an AI generation job."
      },
      "AIJobRetryResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the retry operation was scheduled successfully."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message."
          },
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Target AI job ID."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Updated status after retry schedule (e.g. PENDING)."
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "job_id",
          "status"
        ],
        "title": "AIJobRetryResponse",
        "description": "Response model for an AI job retry operation."
      },
...
      "AIReportResponse": {
        "properties": {
          "executive_summary": {
            "type": "string",
            "title": "Executive Summary"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "explanations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanations"
          },
          "narrative": {
            "type": "string",
            "title": "Narrative"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "execution_time": {
            "type": "number",
            "title": "Execution Time"
          }
        },
        "type": "object",
        "required": [
          "executive_summary",
          "narrative",
          "model",
          "provider",
          "execution_time"
        ],
        "title": "AIReportResponse",
        "description": "AI-generated business insights."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}